feat: CRUD Customer, Business, Order, Job e Schedule

+++ src/main/java/com/agendaai/agendaai/service/JobService.java
@@ -3,7 +3,9 @@ package com.agendaai.agendaai.service;
 import java.util.List;
 import java.util.UUID;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import com.agendaai.agendaai.model.Schedule;
+import com.agendaai.agendaai.repository.ScheduleRepository;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import com.agendaai.agendaai.model.Job;
@@ -11,13 +13,16 @@ import com.agendaai.agendaai.repository.JobRepository;
 
 import jakarta.transaction.Transactional;
 
+@AllArgsConstructor
 @Service
 public class JobService {
-    @Autowired
-    private JobRepository jobRepository;
+
+    private final JobRepository jobRepository;
+    private final ScheduleRepository scheduleRepository;
 
     @Transactional
     public Job createJob(Job job) {
+        scheduleRepository.save(job.getSchedule());
         return jobRepository.save(job);
     }
 
@@ -46,8 +51,27 @@ public class JobService {
         return null;
     }
 
-    public void deleteJob(UUID id) {
-        jobRepository.deleteById(id);
+    @Transactional
+    public Job updateSchedule(UUID jobId, Schedule schedule) {
+        Job existingJob = getJobById(jobId);
+
+        if (existingJob != null) {
+            schedule.setId(existingJob.getSchedule().getId());
+            existingJob.setSchedule(scheduleRepository.save(schedule));
+
+            return jobRepository.save(existingJob);
+        }
+        return null;
+    }
+
+    public boolean deleteJob(UUID id) {
+        Job existingJob = getJobById(id);
+
+        if (existingJob != null) {
+            jobRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
 }
